Refactor Ado Query types.
Move AdoBase to a matching file.
Refactor AdoScalar to derive from AdoBase.
